feat: version 1

+++ app/components/car/modalproduct.js
@@ -7,12 +7,13 @@ import {
     InputAutoComplete,
     TextArea,
     Button,
-    Spinner
+    Spinner,
+    Alert
 } from '../ui';
 
 import { ModalHeader } from '../modal/header';
 import Modal from '../../containers/modal/modal';
-import { MProduct } from './models';
+import {MProduct, Itemorder, Sesion} from './models';
 
 
 const CarModalproduct = {};
@@ -26,7 +27,7 @@ CarModalproduct.vm = function (p) {
             return MProduct.get(id);
         },
         addToCar: (product) => {
-            p.car().push(product);
+            p.order().items_orders().push(product);
         },
         submit(event) {
             if (event) { event.preventDefault(); }
@@ -34,20 +35,32 @@ CarModalproduct.vm = function (p) {
                 return;
             }
         }
-    }
-}
+    };
+};
 
 CarModalproduct.controller = function (p) {
     this.vm = CarModalproduct.vm(p);
     this.vm.refreshProduct(p.product.id()).then(p.product).then(()=>m.redraw());
     this.addToCar = (product) => {
-        product.amount = m.prop(this.vm.amount());
-        product.observations = m.prop(this.vm.observations());
-        this.vm.addToCar(product);
-        Modal.vm.terminate();
-        m.redraw(true);
+
+        if(!Sesion.haveSesionClient()){
+            Modal.vm.open(Alert, {label: 'Porfavor, inicie sesión para poder seleccionar productos'});
+        }else{
+            let params = {}; 
+            params.amount = this.vm.amount();
+            params.name = product.name();
+            params.numberValue = product.numberValue();
+            params.value = product.value();
+            params.observations = this.vm.observations();
+            params.products_id = product.id();
+            params.orders_id = p.order().id();
+            this.vm.addToCar(new Itemorder(params));
+
+            Modal.vm.terminate();
+            m.redraw(true);
+        }
     }; 
-}
+};
 
 CarModalproduct.view = function (c,p) {
     return (
@@ -64,7 +77,7 @@ CarModalproduct.view = function (c,p) {
                     <h4 class="price-product">{p.product.value()}</h4>
                     <p align="justify">{p.product.description()}</p>
                     <div>
-                        Pedir <input type="number" value={c.vm.amount()} oninput={m.withAttr('value', c.vm.amount)} class="pt-input amount-input"/>
+                        Pedir <input type="number" min="1" value={c.vm.amount()} oninput={m.withAttr('value', c.vm.amount)} class="pt-input amount-input"/>
                         <br/>
                         <textarea
                         style="min-height:70px; color: #000000;" 
@@ -77,8 +90,7 @@ CarModalproduct.view = function (c,p) {
                 </div>
             </div>
         </div>
-    )
-}
-
+    );
+};
 
 export default CarModalproduct;
